Revert "Show progress bar when importing videos"
This reverts commit e492abe0ecc2dd9a6ce355c59d5fe59ec6f3387c.

I realised this causes more of a headache than it's worth. Most of the
time is spent constructing the SourcesRow objects. And GTK code needs to
be called in the main thread, since GTK code isn't thread safe (so will
cause a crash when called from another thread). So, there's no point
constructing all these in a function via GLib.idle_add and updating the
progress bar with each row constructed, since the UI won't refresh until
everything is constructed anyway.

Maybe there is a way to solve this. But I guess this is only really
useful when importing a massive number of video sources anyway. And I've
already put off a new release long enough. #42